feat: refactor locale handling by extracting utility functions and improving locale detection
